feat: tests for steps
This commit updates the test-B01_SL_load_single_file.yml workflow file. It changes the name of the workflow to "Test steps" and updates the runs-on value to "ubuntu-latest". Additionally, it adds a step to enable Git LFS and installs pytest and its dependencies. Finally, it updates the run command for the first step and adds a new step to run tests using pytest.