refactor: test_steps.py to run step that takes the longest first
The pytest's `xdist` plug-in seems to assign test functions to workers in the order that they are listed in the file. Because default GitHub runners only have 4 cores, listed fifth `test_step5` -- the step that takes the longest to complete -- does not get picked up immediately. This PR lists this step before the others which should make things run much faster in general.